+++ core/collection.ts
@@ -7,7 +7,7 @@ export interface CollectionOptions {
   label?: string;
   description?: string;
   storage: Storage;
-  fields: Lume.CMS.ResolvedField[];
+  fields: Lume.CMS.ResolvedField;
   url?: string;
   views?: string[] | ((data?: Data) => string[] | undefined);
   documentName?: string | ((changes: Data) => string | undefined);
@@ -30,7 +30,7 @@ export default class Collection {
   label: string;
   description?: string;
   #storage: Storage;
-  #fields: Lume.CMS.ResolvedField[];
+  #fields: Lume.CMS.ResolvedField;
   url?: string;
   views?: string[] | ((data?: Data) => string[] | undefined);
   documentName?: string | ((changes: Data) => string | undefined);