+++ static/components/f-checkbox.js
@@ -1,4 +1,4 @@
-import { oninvalid, view } from "./utils.js";
+import { getFieldName, oninvalid, view } from "./utils.js";
 import { Component } from "./component.js";
 import dom from "dom";
 
@@ -7,8 +7,8 @@ customElements.define(
   class extends Component {
     init() {
       this.classList.add("field");
-      const { schema, value, namePrefix, isNew } = this;
-      const name = `${namePrefix}.${schema.name}`;
+      const { schema, value, isNew } = this;
+      const name = getFieldName(this);
       const id = `field_${name}`;
       const div = dom("div", { class: "field-check" }, this);
       view(this);