+++ static/components/f-datetime.js
@@ -1,4 +1,10 @@
-import { oninvalid, toLocal, updateField, view } from "./utils.js";
+import {
+  getFieldName,
+  oninvalid,
+  toLocal,
+  updateField,
+  view,
+} from "./utils.js";
 import { Component } from "./component.js";
 import dom from "dom";
 
@@ -22,8 +28,8 @@ export class Input extends Component {
 
   init() {
     this.classList.add("field");
-    const { schema, value, namePrefix, isNew } = this;
-    const name = `${namePrefix}.${schema.name}`;
+    const { schema, value, isNew } = this;
+    const name = getFieldName(this);
     const id = `field_${name}`;
 
     view(this);