+++ static/components/f-object.js
@@ -1,5 +1,5 @@
 import { Component } from "./component.js";
-import { view } from "./utils.js";
+import { getFieldName, view } from "./utils.js";
 import dom from "dom";
 
 customElements.define(
@@ -9,7 +9,7 @@ customElements.define(
       this.classList.add("field", "is-group");
       const { schema, value, isNew } = this;
       view(this);
-      const namePrefix = `${this.namePrefix}.${schema.name}`;
+      const namePrefix = getFieldName(this);
       const details = dom("details", {
         class: "accordion",
         "oncms:invalid": () => details.open = true,
@@ -75,6 +75,46 @@ customElements.define(
       this.querySelector(".accordion-description").innerHTML =
         schema.description ??
           "";
+      const items = Array.from(this.querySelector(".fieldset").children);
+
+      for (const field of schema.fields) {
+        items.shift()?.update(field, value?.[field.name]);
+      }
+    }
+  },
+);
+
+/** Version for root mode */
+customElements.define(
+  "f-object-root",
+  class extends Component {
+    init() {
+      const { schema, value, isNew } = this;
+      const namePrefix = getFieldName(this);
+
+      const div = dom("div", {
+        class: "fieldset is-root is-separated",
+      }, this);
+
+      for (const field of schema.fields) {
+        dom(field.tag, {
+          ".schema": field,
+          ".namePrefix": namePrefix,
+          ".isNew": isNew,
+          value: value?.[field.name] ?? null,
+        }, div);
+      }
+    }
+
+    get currentValue() {
+      const values = {};
+      for (const field of this.querySelector(".fieldset").children) {
+        values[field.schema.name] = field.currentValue;
+      }
+      return values;
+    }
+
+    update(schema, value) {
       const items = Array.from(this.querySelector(".fieldset").children);
 
       for (const field of schema.fields) {