+++ static/components/f-select.js
@@ -1,4 +1,10 @@
-import { oninvalid, pushOptions, updateField, view } from "./utils.js";
+import {
+  getFieldName,
+  oninvalid,
+  pushOptions,
+  updateField,
+  view,
+} from "./utils.js";
 import { Component } from "./component.js";
 import dom from "dom";
 
@@ -7,8 +13,8 @@ customElements.define(
   class extends Component {
     init() {
       this.classList.add("field");
-      const { schema, value, namePrefix, isNew } = this;
-      const name = `${namePrefix}.${schema.name}`;
+      const { schema, value, isNew } = this;
+      const name = getFieldName(this);
       const id = `field_${name}`;
 
       view(this);