+++ static/components/f-text.js
@@ -1,5 +1,11 @@
 import { Component } from "./component.js";
-import { oninvalid, pushOptions, updateField, view } from "./utils.js";
+import {
+  getFieldName,
+  oninvalid,
+  pushOptions,
+  updateField,
+  view,
+} from "./utils.js";
 import dom from "dom";
 
 export class Input extends Component {
@@ -9,8 +15,8 @@ export class Input extends Component {
 
   init() {
     this.classList.add("field");
-    const { schema, value, namePrefix, isNew } = this;
-    const name = `${namePrefix}.${schema.name}`;
+    const { schema, value, isNew } = this;
+    const name = getFieldName(this);
     const id = `field_${name}`;
 
     view(this);