+++ static/components/f-textarea.js
@@ -1,5 +1,5 @@
 import { Component } from "./component.js";
-import { oninvalid, updateField, view } from "./utils.js";
+import { getFieldName, oninvalid, updateField, view } from "./utils.js";
 import dom from "dom";
 
 customElements.define(
@@ -7,8 +7,8 @@ customElements.define(
   class extends Component {
     init() {
       this.classList.add("field");
-      const { schema, value, namePrefix, isNew } = this;
-      const name = `${namePrefix}.${schema.name}`;
+      const { schema, value, isNew } = this;
+      const name = getFieldName(this);
       const id = `field_${name}`;
 
       view(this);